Split dataflow optimize from assemble.
